python/requirements.txt
mouse===0.7.1
pyserial===3.5